@@ -1,11 +1,11 @@
 {
-  "name": "4MoBeers",
+  "name": "4mobeers",
   "version": "0.1.0",
-  "description": "4MoBeers",
+  "description": "4MoBeers DAO Inc.",
   "private": true,
   "scripts": {
     "dev": "next dev",
-    "build": "next build && next export",
+    "build": "next build && next export ",
     "start": "next start",
     "lint": "next lint",
     "export": "next export",
@@ -18,7 +18,7 @@
   "keywords": [
     "nextjs",
     "tailwindcss",
-    "free"
+    "rainbowkit"
   ],
   "author": "4MoBeers DAO Inc",
   "license": "ISC",
@@ -32,52 +32,42 @@
     "@fortawesome/react-fontawesome": "^0.2.0",
     "@headlessui/react": "^1.2.0",
     "@heroicons/react": "^1.0.1",
+    "@lottiefiles/react-lottie-player": "^3.5.3",
     "@nextui-org/react": "^1.0.0-beta.10",
-    "@solana/wallet-adapter-react-ui": "^0.9.27",
-    "@supabase/auth-helpers-nextjs": "^0.5.2",
+    "@rainbow-me/rainbowkit": "^0.8.1",
+    "@rainbow-me/rainbowkit-siwe-next-auth": "^0.1.5",
+    "@supabase/auth-helpers-nextjs": "^0.5.4",
     "@supabase/auth-helpers-react": "^0.3.1",
     "@supabase/auth-ui-react": "^0.2.6",
     "@supabase/supabase-js": "^2.2.2",
     "@tailwindcss/aspect-ratio": "^0.2.1",
-    "@thirdweb-dev/auth": "^2.0.39",
-    "@thirdweb-dev/react": "^3.6.8",
-    "@thirdweb-dev/sdk": "^3.6.8",
-    "@uauth/web3-react": "^2.4.1-rc.0",
-    "@walletconnect/ethereum-provider": "^1.8.0",
-    "@web3-react/core": "^8.0.35-beta.0",
-    "@web3-react/metamask": "^8.0.28-beta.0",
-    "@web3-react/walletconnect": "^8.0.36-beta.0",
-    "aos": "^2.3.4",
     "autoprefixer": "^10.0.4",
-    "ethers": "^5.7.2",
-    "next": "^13.1.2",
+    "ethers": "^6.3.0",
+    "next": "^13.3.0",
+    "next-auth": "^4.22.0",
     "next-themes": "^0.2.1",
     "postcss": "^8.1.10",
     "react": "^18.2.0",
-    "react-aria": "^3.22.0",
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.39.2",
-    "rimraf": "^4.1.1",
-    "sass": "^1.57.1",
-    "tailwind": "^2.2.0",
-    "tailwindcss": "^3.2.4",
-    "typescript": "^4.8.4"
+    "siwe": "^1.1.6",
+    "tailwindcss": "^3.3.1",
+    "wagmi": "^0.9.6"
+  },
+  "peerDependencies": {
+    "siwe": "^2.1.3"
   },
   "devDependencies": {
-    "@babel/core": "^7.0.0",
-    "@types/react": "latest",
-    "@typescript-eslint/parser": "latest",
-    "babel-plugin-macros": "^3.1.0",
-    "babel-plugin-twin": "^1.0.2",
+    "ethers": "5.5.1",
+    "@tailwindcss/forms": "^0.5.3",
     "eslint": "8.27.0",
     "eslint-config-next": "13.0.2",
-    "tailwindcss": "^3.2.4",
-    "twin.macro": "^3.0.0"
+    "prettier": "^2.8.7",
+    "prettier-plugin-tailwindcss": "^0.2.7"
   },
-  "babelMacros": {
-    "twin": {
-      "preset": "emotion",
-      "includeClassNames": true
-    }
+  "browser": {
+    "fs": false,
+    "os": false,
+    "path": false
   }
 }
