@@ -1,3 +1,4 @@
+'use client';
 import { useEffect } from "react";
 import { getCalApi } from "@calcom/embed-react";
 
@@ -6,16 +7,17 @@ export default function MyCal() {
     (async function () {
       const cal = await getCalApi();
       cal("floatingButton", {
-        calLink: "4mobeers/30min",
-        buttonText: "Beer Me",
-        buttonColor: "#7e22ce",
+        calLink: "4mobeers/15min",
+        buttonText: "🍻",
+        buttonColor: "#a21caf",
         buttonTextColor: "#f4f4f5",
         buttonPosition: "bottom-left",
+        hideButtonIcon: true,
       });
       cal("ui", {
         theme: "dark",
         styles: {
-          branding: { brandColor: "#ea580c" },
+          branding: { brandColor: "#a21caf" },
         },
       });
     })();
