@@ -3,7 +3,7 @@ import React from "react";
 export default function Container(props) {
   return (
     <div
-      className={`container p-8 mx-auto ${
+      className={`container py-4 mx-auto ${
         props.className ? props.className : ""
       }`}
     >
