@@ -1,43 +1,46 @@
-import { useState, useEffect } from "react";
-import { useRouter } from "next/router";
-import Image from "next/image";
-import Avatar from "./avatar";
-
-import { useUser, useAddress } from "@thirdweb-dev/react";
-
-//import { useSession, useSupabaseClient } from "@supabase/auth-helpers-react";
-//import { useSupabaseClient } from "@supabase/supabase-js";
-import { supabase } from "../../utils/supabase-clients";
+import prettier from "prettier";
+import React, { useState, useEffect } from "react";
+import { useSupabaseClient } from "@supabase/auth-helpers-react";
+import { useAccount, useConnect } from "wagmi";
+import {
+  Button,
+  Spacer,
+  Input,
+  useInput,
+  Text,
+  Loading,
+} from "@nextui-org/react";
+import { Player } from "@lottiefiles/react-lottie-player";
 
 export default function Profile({ address }) {
-  //const router = useRouter();
-  //const supabase = useSupabaseClient();
+  const supabase = useSupabaseClient();
+  const { connector: activeConnector } = useAccount();
+  const provider = activeConnector.name;
+  //const provider = { walletprovider };
 
-  const user = useUser();
-  const connectedaddress = useAddress();
   const [loading, setLoading] = useState(true);
-  const [walletaddress, setWalletAddress] = useState(null);
+
+  //const [walletaddress, setWalletAddress] = useState(null);
   const [uuid, setUUID] = useState(null);
-  const [email, setEmail] = useState(null);
   const [beername, setBeername] = useState(null);
-  const [fullname, setFullname] = useState(null);
   const [beertitle, setBeertitle] = useState(null);
-  const [avatar_url, setAvatarUrl] = useState(null);
+  const [email, setEmail] = useState(null);
+  //const [provider, setProvider] = useState(null);
+  //const [fullname, setFullname] = useState(null);
+  const [isSuccess, setIsSuccess] = useState(false);
 
   useEffect(() => {
     getProfile();
-  }, [address]);
+  });
 
-  async function getProfile(address) {
+  async function getProfile() {
     try {
       setLoading(true);
 
       let { data, error, status } = await supabase
         .from("users")
-        //.select("walletaddress, beername, fullname, email, beertitle, avatar_url")
         .select("*")
-        .eq("walletaddress", connectedaddress)
-        //.eq("id", user.id)
+        .eq("walletaddress", address)
         .single();
 
       if (error && status !== 406) {
@@ -46,175 +49,173 @@ export default function Profile({ address }) {
 
       if (data) {
         setUUID(data.id);
-        setWalletAddress(data.walletaddress);
-        setFullname(data.fullname);
         setEmail(data.email);
         setBeername(data.beername);
         setBeertitle(data.beertitle);
-        setAvatarUrl(data.avatar_url);
+        setProvider(walletprovider);
       }
     } catch (error) {
-      alert("Error loading user data!");
+      //alert("Error loading user data!");
       console.log(error);
     } finally {
       setLoading(false);
     }
   }
 
-  async function updateProfile({
-    fullname,
-    email,
-    beername,
-    beertitle,
-    avatar_url,
-  }) {
+  async function updateProfile({ uuid, beername, beertitle, email, provider }) {
     try {
       setLoading(true);
 
       const updates = {
-        id: uuid,
-        fullname,
-        email,
+        //id: uuid,
+        walletaddress: address,
         beername,
         beertitle,
-        avatar_url,
+        email,
+        provider,
         updated_at: new Date().toISOString(),
       };
 
       let { error } = await supabase.from("users").upsert(updates);
-      //.eq("walletaddress", connectedaddress);
-
-      //.eq("id", id);
+      //.eq("walletaddress", address);
+      {
+        /*} let { error } = await supabase
+        .from("users")
+        .update({
+          // id: "d18baef1-f294-4b54-bdc4-7fd76b12758b",
+          //walletaddress: "0x283A6f1D0974e10c20047480dA08FB2874911504",
+          beername: "HELLO",
+          beertitle: "HELLO", //beertitle,
+          email: "HELLO@hello.com", //email,
+          fullname: "HELLO",
+          provider: "HELLO", //walletprovider,
+          updated_at: new Date().toISOString(),
+        })
+        .eq("id", "d18baef1-f294-4b54-bdc4-7fd76b12758b");
+        */
+      }
       if (error) throw error;
       alert("Profile updated!");
     } catch (error) {
       alert("Error updating the data!");
       console.log(error);
     } finally {
+      setIsSuccess(true);
       setLoading(false);
     }
   }
-  return (
-    <>
-      <div className="form-widget align-center flex-grow h-full p-6 overflow-auto">
-        {/* <div className="p-8">
-          {avatar_url ? (
-            <Image
-              src={avatar_url}
-              alt="Avatar"
-              className="rounded-full mx-auto"
-              width={150}
-              height={150}
-            />
-          ) : (
-            <div className="avatar no-image" width={150} height={150}>
-              <Image
-                src="https://storageapi.fleek.co/9ec37732-832c-4af1-a4fa-cc87c28c6dae-bucket/assets/img/memberprofile.png"
-                alt="Avatar"
-                className="rounded-half mx-auto"
-                width={150}
-                height={150}
-              />
-            </div>
-          )}
-        </div>*/}
-        {/*
-        <div className="py-4 align-center mb-4 bg-trueZinc-900">
-             <Avatar
-            uid={user.id}
-            url={avatar_url}
-            size={150}
-            onUpload={(url) => {
-              setAvatarUrl(url);
-              updateProfile({ beername, avatar_url: url });
-            }}
-            
-          />
-          
-        </div>
-*/}
-        <div className="mb-4 font-bold">
-          <label htmlFor="walletaddress">Wallet Address</label>
-          <input
-            id="walletaddress"
-            type="text"
-            value={connectedaddress || ""}
-            disabled
-            className="w-full text-truePink-500 rounded-md p-2 mt-4"
-          />
-        </div>
-        {/*<div className="mb-4">
-          <label htmlFor="id">UserID</label>
-          <input
-            id="uuid"
-            type="text"
-            value={uuid || ""}
-            disabled
-            className="w-full text-truePink-500 rounded-md p-2 mt-4"
-          />
-        </div>*/}
-        <div className="mb-4 font-bold">
-          <label htmlFor="beername">Beer Name (Required)</label>
-          <input
-            id="beername"
-            type="text"
-            value={beername || ""}
-            onChange={(e) => setBeername(e.target.value)}
-            placeholder="Drink A Lot"
-            className="w-full rounded-md p-2 mt-4"
-          />
-        </div>
-        <div className="mb-4 font-bold">
-          <label htmlFor="beertitle">Beer Title (Required)</label>
-          <input
-            id="beertitle"
-            type="beertitle"
-            value={beertitle || ""}
-            onChange={(e) => setBeertitle(e.target.value)}
-            placeholder="Beer King"
-            className="w-full rounded-md p-2 mt-4"
-          />
-        </div>
-        <div className="mb-4 font-bold">
-          <label htmlFor="email">Email (Required)</label>
-          <input
-            id="email"
-            type="text"
-            value={email || ""}
-            onChange={(e) => setEmail(e.target.value)}
-            placeholder="beerking@email.com"
-            className="w-full rounded-md p-2 mt-4"
-          />
-        </div>
-        <div className="mb-4 font-bold">
-          <label htmlFor="fullname">Alias/Full Name (Optional)</label>
-          <input
-            id="fullname"
-            type="text"
-            value={fullname || ""}
-            onChange={(e) => setFullname(e.target.value)}
-            className="w-full rounded-md p-2 mt-4"
-          />
-        </div>
 
-        <div className="align-center mb-4">
-          <button
-            className="button primary block bg-truePurple-900 rounded-md mt-8 p-2 w-full text-trueZinc-100"
-            onClick={() =>
-              updateProfile({
-                fullname,
-                email,
-                beername,
-                beertitle,
-                avatar_url,
-              })
-            }
-            disabled={loading}
-          >
-            {loading ? "Loading ..." : "Update"}
-          </button>
-        </div>
+  return (
+    <div className="form-widget align-center overflow-none h-full flex-grow">
+      <div className="mb-4 font-bold text-truePink-600">
+        {loading ? (
+          <>
+            <Text h1>
+              LOADING PROFILE ... <Loading type="points" size="sm" />
+              <Spacer />
+            </Text>
+          </>
+        ) : (
+          <>
+            <Input
+              id="uuid"
+              label="Profile ID"
+              type="text"
+              value={uuid}
+              disabled
+              fullWidth
+              color="secondary"
+              textColor="$black700"
+            />
+          </>
+        )}
       </div>
-    </>
+
+      <div className="mb-4 font-bold text-truePink-600">
+        <Input
+          id="beername"
+          label="Beer Name"
+          type="text"
+          value={beername || ""}
+          onChange={(e) => setBeername(e.target.value)}
+          color="secondary"
+          fullWidth
+          clearable
+          bordered
+        />
+      </div>
+      <div className="mb-4 font-bold text-truePink-600">
+        <Input
+          id="beertitle"
+          label="Beer Title"
+          type="text"
+          value={beertitle || ""}
+          onChange={(e) => setBeertitle(e.target.value)}
+          color="secondary"
+          fullWidth
+          clearable
+          bordered
+          className=""
+        />
+      </div>
+      <div className="mb-4 font-bold text-truePink-600">
+        <Input
+          id="email"
+          label="Email"
+          type="email"
+          value={email || ""}
+          onChange={(e) => setEmail(e.target.value)}
+          color="secondary"
+          fullWidth
+          clearable
+          bordered
+          className="autofill:bg-current"
+        />
+      </div>
+      <div className="align-center mb-4 mt-4">
+        <Button
+          fullWidth
+          animated
+          ripple
+          shadow
+          className="button primary block w-full bg-trueFushia-700 uppercase text-trueZinc-100"
+          onPress={() =>
+            updateProfile({ uuid, beername, beertitle, email, provider })
+          }
+          //disabled={loading}
+        >
+          Update
+        </Button>
+      </div>
+      {isSuccess ? (
+        <>
+          <div className="flex h-full flex-col items-center justify-center rounded-md text-center text-trueZinc-100">
+            <Player
+              autoplay
+              loop
+              src="https://storage.fleek.zone/ead18a13-670a-4701-91a7-06a767c3dcd5-bucket/assets/lottie/drunkbeerbottlegreen.json"
+              style={{ height: "100px", width: "100px" }}
+            ></Player>
+            <h2 className="py-5 text-xl uppercase text-trueEmerald-500">
+              Cheers!!
+            </h2>
+          </div>
+        </>
+      ) : (
+        <>
+          <div className="flex h-full flex-col items-center justify-center rounded-md text-center text-trueZinc-100">
+            <Player
+              autoplay
+              loop
+              src="https://storage.fleek.zone/ead18a13-670a-4701-91a7-06a767c3dcd5-bucket/assets/lottie/drunkbeerbottlered.json"
+              style={{ height: "100px", width: "100px" }}
+            ></Player>
+            <h2 className="py-5 text-xl uppercase text-trueEmerald-500">
+              Oops!! We spilled your drink.
+            </h2>
+          </div>
+        </>
+      )}
+    </div>
   );
 }
