@@ -1,12 +1,16 @@
+"use client";
 import prettier from "prettier";
 import React, { useState } from "react";
 import { useForm } from "react-hook-form";
-import Link from "next/link";
-import Container from "./container";
 import { useSupabaseClient } from "@supabase/auth-helpers-react";
 import { useRouter } from "next/router";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { faArrowCircleLeft } from "@fortawesome/free-solid-svg-icons";
+import { faDiscord } from "@fortawesome/free-brands-svg-icons";
+import { Button, Link } from "@nextui-org/react";
+import { Player } from "@lottiefiles/react-lottie-player";
 
-export default function Test() {
+export default function ParterSignup() {
   const {
     register,
     handleSubmit,
@@ -40,7 +44,16 @@ export default function Test() {
           updated_at: new Date().toISOString(),
         };
 
-        let { error } = await supabase.from("partners").insert(registerPartner);
+        let { error } = await supabase
+          .from("partners")
+          .insert(
+            companyname,
+            email,
+            clientname,
+            contactnumber,
+            website,
+            businesstype
+          );
         if (error) throw error;
         //alert("Whitelist updated!");
       } catch (error) {
@@ -56,40 +69,40 @@ export default function Test() {
     };
 
   return (
-    <div className="container mx-auto flex w-10/12 w-full flex-wrap place-content-center p-8">
-      <div className="left-0 flex h-full min-h-[250px]  w-full flex-col overflow-hidden bg-trueBlue-300 dark:bg-trueZinc-900 sm:h-[600px] sm:max-h-[calc(100vh-120px)] sm:w-[350px]">
-        <div className="h-full flex-grow overflow-auto p-6">
+    <div className="mx-auto flex w-full flex-wrap place-content-center">
+      <div className="left-0 flex  h-full w-full flex-col overflow-hidden ">
+        <div className="h-full flex-grow overflow-auto">
           {!isSubmitSuccessful && (
             <form onSubmit={handleSubmit(onSubmit)} noValidate>
-              <div className="mb-4 px-4">
+              <div id="companyname" className="mb-4 px-4">
                 <label
-                  htmlFor="companyName"
+                  htmlFor="companyname"
                   className="mb-2 block text-sm text-truePink-600 dark:text-truePink-400"
                 >
                   Company Name
                 </label>
                 <input
                   type="text"
-                  id="companyName"
+                  id="companyname"
                   placeholder="Greatest Brew Co."
-                  {...register("companyName", {
+                  {...register("companyname", {
                     required: "Company name is required",
                     maxLength: 80,
                   })}
                   className={`w-full rounded-md border border-trueZinc-300 bg-trueZinc-100 px-3 py-2 placeholder-trueZinc-300 focus:outline-none focus:ring   ${
-                    errors.companyName
-                      ? "border-red-600 ring-red-100 focus:border-red-600"
+                    errors.companyname
+                      ? "border-red-600 ring-red-700 focus:border-red-600"
                       : "border-trueZinc-300 ring-trueFushia-100 focus:border-trueFushia-600"
                   }`}
                 />
-                {errors.companyName && (
+                {errors.companyname && (
                   <div className="invalid-feedback mt-1 text-sm text-red-400">
-                    {errors.companyName.message}
+                    {errors.companyname.message}
                   </div>
                 )}
               </div>
 
-              <div className="mb-4 px-4">
+              <div id="email" className="mb-4 px-4">
                 <label
                   htmlFor="Email"
                   className="mb-2 block text-sm text-truePink-600 dark:text-truePink-400"
@@ -99,7 +112,7 @@ export default function Test() {
                 <input
                   type="email"
                   id="Email"
-                  {...register("Email", {
+                  {...register("email", {
                     required: "Enter your email",
                     pattern: {
                       value: /^\S+@\S+$/i,
@@ -108,20 +121,20 @@ export default function Test() {
                   })}
                   placeholder="brewbuddy@company.com"
                   className={`w-full rounded-md border border-trueZinc-300 bg-trueZinc-100 px-3 py-2 placeholder-trueZinc-300 focus:outline-none focus:ring   ${
-                    errors.Email
-                      ? "border-red-600 ring-red-100 focus:border-red-600"
+                    errors.email
+                      ? "border-red-600 ring-red-700 focus:border-red-600"
                       : "border-trueZinc-300 ring-trueFushia-100 focus:border-trueFushia-600"
                   }`}
                 />
 
-                {errors.Email && (
+                {errors.email && (
                   <div className="invalid-feedback mt-1 text-sm text-red-400">
-                    {errors.Email.message}
+                    {errors.email.message}
                   </div>
                 )}
               </div>
 
-              <div className="mb-4 px-4">
+              <div id="clientname" className="mb-4 px-4">
                 <label
                   htmlFor="clientName"
                   className="mb-2 block text-sm text-truePink-600 dark:text-truePink-400"
@@ -130,57 +143,60 @@ export default function Test() {
                 </label>
                 <input
                   type="text"
-                  id="clientName"
+                  id="clientname"
                   placeholder="Brew Buddy"
-                  {...register("clientName", {
+                  {...register("clientname", {
                     required: "Your full name is required",
                     maxLength: 80,
                   })}
                   className={`w-full rounded-md border border-trueZinc-300 bg-trueZinc-100 px-3 py-2 placeholder-trueZinc-300 focus:outline-none focus:ring   ${
-                    errors.clientName
-                      ? "border-red-600 ring-red-100 focus:border-red-600"
+                    errors.clientname
+                      ? "border-red-600 ring-red-700 focus:border-red-600"
                       : "border-trueZinc-300 ring-trueFushia-100 focus:border-trueFushia-600"
                   }`}
                 />
-                {errors.clientName && (
+                {errors.clientname && (
                   <div className="invalid-feedback mt-1 text-sm text-red-400">
-                    {errors.clientName.message}
+                    {errors.clientname.message}
                   </div>
                 )}
               </div>
 
-              <div className="mb-4 px-4">
+              <div id="contactnumber" className="mb-4 px-4">
                 <label
-                  htmlFor="clientName"
+                  htmlFor="contactnumber"
                   className="mb-2 block text-sm text-truePink-600 dark:text-truePink-400"
                 >
                   Contact Number
                 </label>
                 <input
-                  type="text"
-                  id="contactNumber"
+                  type="tel"
+                  id="contactnumber"
                   placeholder="000-000-0000"
-                  {...register("contactNumber", {
+                  {...register("contactnumber", {
                     required:
                       "Please enter your contact number e.g.000-000-0000 ",
-                    maxLength: 80,
+
+                    min: 12,
+                    maxLength: 12,
+                    pattern: /^\S+-\S+-\S/i,
                   })}
                   className={`w-full rounded-md border border-trueZinc-300 bg-trueZinc-100 px-3 py-2 placeholder-trueZinc-300 focus:outline-none focus:ring   ${
-                    errors.contactNumber
-                      ? "border-red-600 ring-red-100 focus:border-red-600"
+                    errors.contactnumber
+                      ? "border-red-600 ring-red-700 focus:border-red-600"
                       : "border-trueZinc-300 ring-trueFushia-100 focus:border-trueFushia-600"
                   }`}
                 />
-                {errors.contactNumber && (
+                {errors.contactnumber && (
                   <div className="invalid-feedback mt-1 text-sm text-red-400">
-                    {errors.contactNumber.message}
+                    {errors.contactnumber.message}
                   </div>
                 )}
               </div>
 
-              <div className="mb-4 px-4">
+              <div id="website" className="mb-4 px-4">
                 <label
-                  htmlFor="clientName"
+                  htmlFor="website"
                   className="mb-2 block text-sm text-truePink-600 dark:text-truePink-400"
                 >
                   Website
@@ -188,14 +204,15 @@ export default function Test() {
                 <input
                   type="url"
                   id="website"
-                  placeholder="000-000-0000"
+                  placeholder="greatestbrew.co"
                   {...register("website", {
                     required: "Please enter the company's website URL",
                     maxLength: 80,
+                    pattern: /^\S+.\S/i,
                   })}
                   className={`w-full rounded-md border border-trueZinc-300 bg-trueZinc-100 px-3 py-2 placeholder-trueZinc-300 focus:outline-none focus:ring   ${
                     errors.website
-                      ? "border-red-600 ring-red-100 focus:border-red-600"
+                      ? "border-red-600 ring-red-700 focus:border-red-600"
                       : "border-trueZinc-300 ring-trueFushia-100 focus:border-trueFushia-600"
                   }`}
                 />
@@ -206,9 +223,9 @@ export default function Test() {
                 )}
               </div>
 
-              <div className="mb-4 px-4">
+              <div id="businesstype" className="mb-4 px-4">
                 <label
-                  htmlFor="clientName"
+                  htmlFor="businesstype"
                   className="mb-2 block text-sm text-truePink-600 dark:text-truePink-400"
                 >
                   Website
@@ -219,14 +236,14 @@ export default function Test() {
                   })}
                   className={`w-full rounded-md border border-trueZinc-300 bg-trueZinc-100 px-3 py-2 placeholder-trueZinc-300 focus:outline-none focus:ring   ${
                     errors.businesstype
-                      ? "border-red-600 ring-red-100 focus:border-red-600"
+                      ? "border-red-600 ring-red-700 focus:border-red-600"
                       : "border-trueZinc-300 ring-trueFushia-100 focus:border-trueFushia-600"
                   }`}
                 >
                   <option value="default">Select An Option</option>
                   <option value="Bar">Bar</option>
-
                   <option value="Brewery">Brewery</option>
+                  <option value="Club">Club</option>
                   <option value="Distillery">Distillery</option>
                   <option value="Food Truck">Food Truck</option>
                   <option value="Pub">Pub</option>
@@ -280,71 +297,63 @@ export default function Test() {
           {isSubmitSuccessful && isSuccess && (
             <>
               <div className="flex h-full flex-col items-center justify-center rounded-md text-center text-trueZinc-100">
-                <svg
-                  width="60"
-                  height="60"
-                  className="text-trueFushia-900 dark:text-truePink-400"
-                  viewBox="0 0 100 100"
-                  fill="none"
-                  xmlns="http://www.w3.org/2000/svg"
-                >
-                  <path
-                    d="M26.6666 50L46.6666 66.6667L73.3333 33.3333M50 96.6667C43.8716 96.6667 37.8033 95.4596 32.1414 93.1144C26.4796 90.7692 21.3351 87.3317 17.0017 82.9983C12.6683 78.6649 9.23082 73.5204 6.8856 67.8586C4.54038 62.1967 3.33331 56.1283 3.33331 50C3.33331 43.8716 4.54038 37.8033 6.8856 32.1414C9.23082 26.4796 12.6683 21.3351 17.0017 17.0017C21.3351 12.6683 26.4796 9.23084 32.1414 6.88562C37.8033 4.5404 43.8716 3.33333 50 3.33333C62.3767 3.33333 74.2466 8.24998 82.9983 17.0017C91.75 25.7534 96.6666 37.6232 96.6666 50C96.6666 62.3768 91.75 74.2466 82.9983 82.9983C74.2466 91.75 62.3767 96.6667 50 96.6667Z"
-                    stroke="currentColor"
-                    strokeWidth="3"
-                  />
-                </svg>
+                <Player
+                  autoplay
+                  loop
+                  src="https://storage.fleek.zone/ead18a13-670a-4701-91a7-06a767c3dcd5-bucket/assets/lottie/drunkbeerbottlegreen.json"
+                  style={{ height: "100px", width: "100px" }}
+                ></Player>
                 <h3 className="py-5 text-xl text-trueEmerald-500">
-                  Whitelist request sent successfully
+                  Partner request sent successfully!
                 </h3>
-                <h2 className="py-5 text-xl text-trueZinc-700">
-                  <Link
-                    href="https://discord.gg/d6KnD7V2gs"
-                    alt="Discord"
-                    target="_blank"
-                    className="text-trueZinc-700"
-                  >
-                    Join Our Discord Channel
-                  </Link>
+                <h2 className="py-2 text-xl text-trueZinc-700">
+                  Join Our Partners Discord Channel
                 </h2>
 
-                <button
-                  className="mt-6 rounded-md bg-trueFushia-900 px-4 py-2 text-trueZinc-100 focus:outline-none"
-                  onClick={() => reset()}
+                <Link
+                  href="https://discord.gg/VpUvAbaenh"
+                  alt="Discord"
+                  target="_blank"
+                  rel="noreferrer"
+                  className="text-trueZinc-700"
                 >
-                  Go back
-                </button>
+                  <Button
+                    className="button primary mx-auto block w-full w-full rounded-md bg-trueFushia-900 p-2 uppercase text-trueZinc-100"
+                    auto
+                    shadow
+                    css={{ color: "#f4f4f5", background: "#581c87" }}
+                    icon={<FontAwesomeIcon icon={faDiscord} />}
+                  >
+                    Discord
+                  </Button>
+                </Link>
               </div>
             </>
           )}
 
           {isSubmitSuccessful && !isSuccess && (
             <div className="flex h-full flex-col items-center justify-center rounded-md text-center text-trueZinc-100">
-              <svg
-                width="60"
-                height="60"
-                viewBox="0 0 97 97"
-                className="text-red-400"
-                fill="none"
-                xmlns="http://www.w3.org/2000/svg"
-              >
-                <path
-                  d="M27.9995 69C43.6205 53.379 52.3786 44.621 67.9995 29M26.8077 29L67.9995 69M48.2189 95C42.0906 95 36.0222 93.7929 30.3604 91.4477C24.6985 89.1025 19.554 85.6651 15.2206 81.3316C10.8872 76.9982 7.44975 71.8538 5.10454 66.1919C2.75932 60.53 1.55225 54.4617 1.55225 48.3333C1.55225 42.205 2.75932 36.1366 5.10454 30.4748C7.44975 24.8129 10.8872 19.6684 15.2206 15.335C19.554 11.0016 24.6985 7.56418 30.3604 5.21896C36.0222 2.87374 42.0906 1.66667 48.2189 1.66667C60.5957 1.66667 72.4655 6.58333 81.2172 15.335C89.9689 24.0867 94.8856 35.9566 94.8856 48.3333C94.8856 60.7101 89.9689 72.58 81.2172 81.3316C72.4655 90.0833 60.5957 95 48.2189 95Z"
-                  stroke="CurrentColor"
-                  strokeWidth="3"
-                />
-              </svg>
+              <Player
+                autoplay
+                loop
+                src="https://storage.fleek.zone/ead18a13-670a-4701-91a7-06a767c3dcd5-bucket/assets/lottie/drunkbeerbottlered.json"
+                style={{ height: "100px", width: "100px" }}
+              ></Player>
 
               <h3 className="py-7 text-xl text-red-400">
-                Oops, Something went wrong!
+                Oops, We Spilled Your Drink! PLease Try Again.
               </h3>
 
-              <button
-                className="mt-6 text-trueFushia-600 focus:outline-none"
+              <Button
+                className="button primary mx-auto block w-full w-full rounded-md bg-trueFushia-900 p-2 uppercase text-trueZinc-100"
+                auto
+                shadow
+                css={{ color: "#f4f4f5", background: "#581c87" }}
                 onClick={() => reset()}
               >
-                Go back
-              </button>
+                <FontAwesomeIcon icon={faArrowCircleLeft} className="mr-3" />{" "}
+                Back
+              </Button>
             </div>
           )}
         </div>
