@@ -1,6 +1,9 @@
+"use client";
+import prettier from "prettier";
 import React, { useState } from "react";
 import { useForm, useWatch } from "react-hook-form";
 import { Disclosure, Transition } from "@headlessui/react";
+import { Player } from "@lottiefiles/react-lottie-player";
 import Link from "next/link";
 
 export default function PopupWidget() {
@@ -16,7 +19,11 @@ export default function PopupWidget() {
   const [isSuccess, setIsSuccess] = useState(false);
   const [Message, setMessage] = useState("");
 
-  const userName = useWatch({ control, name: "name", defaultValue: "Someone" });
+  const userName = useWatch({
+    control,
+    name: "name",
+    defaultValue: "Beer Guy",
+  });
 
   const onSubmit = async (data, e) => {
     console.log(data);
@@ -52,7 +59,7 @@ export default function PopupWidget() {
       <Disclosure>
         {({ open }) => (
           <>
-            <Disclosure.Button className="fixed z-40 flex items-center justify-center transition duration-300 bg-truePurple-700 rounded-full shadow-lg right-5 bottom-5 w-14 h-14 focus:outline-none hover:bg-trueSky-500 focus:bg-truePurple-600 ease">
+            <Disclosure.Button className="ease fixed bottom-5 right-5 z-40 flex h-14 w-14 items-center justify-center rounded-full bg-trueFushia-700 shadow-lg transition duration-300 hover:bg-truePink-600 focus:bg-trueFushia-600 focus:outline-none">
               <span className="sr-only">Open Contact form Widget</span>
               <Transition
                 show={!open}
@@ -60,11 +67,11 @@ export default function PopupWidget() {
                 enterFrom="opacity-0 -rotate-45 scale-75"
                 leave="transition duration-100 transform ease"
                 leaveTo="opacity-0 -rotate-45"
-                className="absolute w-6 h-6 text-trueZinc-100"
+                className="absolute h-6 w-6 text-trueZinc-100"
               >
                 <svg
                   xmlns="http://www.w3.org/2000/svg"
-                  className="w-6 h-6"
+                  className="h-6 w-6"
                   width="24"
                   height="24"
                   viewBox="0 0 24 24"
@@ -84,11 +91,11 @@ export default function PopupWidget() {
                 enterFrom="opacity-0 rotate-45 scale-75"
                 leave="transition duration-100 transform ease"
                 leaveTo="opacity-0 rotate-45"
-                className="absolute w-6 h-6 text-trueZinc-100"
+                className="absolute h-6 w-6 text-trueZinc-100"
               >
                 <svg
                   xmlns="http://www.w3.org/2000/svg"
-                  className="w-6 h-6"
+                  className="h-6 w-6"
                   width="24"
                   height="24"
                   viewBox="0 0 24 24"
@@ -104,22 +111,29 @@ export default function PopupWidget() {
               </Transition>
             </Disclosure.Button>
             <Transition
-              className="fixed  z-50 bottom-[100px] top-0 right-0  left-0 sm:top-auto sm:right-5 sm:left-auto"
+              className="fixed  bottom-[100px] left-0 right-0 top-0  z-50 sm:left-auto sm:right-5 sm:top-auto"
               enter="transition duration-200 transform ease"
               enterFrom="opacity-0 translate-y-5"
               leave="transition duration-200 transform ease"
               leaveTo="opacity-0 translate-y-5"
             >
-              <Disclosure.Panel className=" flex flex-col  overflow-hidden left-0 h-full w-full sm:w-[350px] min-h-[250px] sm:h-[600px] border border-trueZinc-300 dark:border-trueZinc-800 bg-trueZinc-100 shadow-2xl rounded-md sm:max-h-[calc(100vh-120px)]">
-                <div className="flex flex-col items-center justify-center h-32 p-5 bg-truePurple-600">
+              <Disclosure.Panel className=" left-0 flex  h-full min-h-[250px] w-full flex-col overflow-hidden rounded-md border border-trueZinc-300 bg-trueZinc-100 shadow-2xl dark:border-trueZinc-800 sm:h-[600px] sm:max-h-[calc(100vh-120px)] sm:w-[350px]">
+                <div className="flex flex-col items-center justify-center bg-trueFushia-700 p-5">
                   <h3 className="text-lg text-trueZinc-100">
                     How can we help?
                   </h3>
+                  <LottiePlayer
+                    autoplay
+                    loop
+                    src="https://storage.fleek.zone/ead18a13-670a-4701-91a7-06a767c3dcd5-bucket/assets/lottie/drunkenowl.json"
+                    style={{ height: "150px", width: "150px" }}
+                  ></LottiePlayer>
+
                   <p className="text-trueZinc-100 opacity-50">
                     We usually respond in a few hours
                   </p>
                 </div>
-                <div className="flex-grow h-full p-6 overflow-auto bg-trueZinc-50">
+                <div className="h-full flex-grow overflow-auto bg-trueZinc-50 p-6">
                   {!isSubmitSuccessful && (
                     <form onSubmit={handleSubmit(onSubmit)} noValidate>
                       <input
@@ -147,7 +161,7 @@ export default function PopupWidget() {
                       <div className="mb-4">
                         <label
                           htmlFor="full_name"
-                          className="block mb-2 text-sm text-trueZinc-600 dark:text-trueZinc-400"
+                          className="mb-2 block text-sm text-trueZinc-700"
                         >
                           Full Name
                         </label>
@@ -159,14 +173,14 @@ export default function PopupWidget() {
                             required: "Full name is required",
                             maxLength: 80,
                           })}
-                          className={`w-full px-3 py-2 placeholder-trueZinc-300 bg-trueZinc-100 border border-trueZinc-300 rounded-md focus:outline-none focus:ring   ${
+                          className={`w-full rounded-md border border-trueZinc-300 bg-trueZinc-100 px-3 py-2 placeholder-trueZinc-300 focus:outline-none focus:ring   ${
                             errors.name
-                              ? "border-red-600 focus:border-red-600 ring-red-100"
-                              : "border-trueZinc-300 focus:border-truePurple-600 ring-truePurple-100"
+                              ? "border-red-600 ring-red-100 focus:border-red-600"
+                              : "border-trueZinc-300 ring-trueFushia-100 focus:border-trueFushia-600"
                           }`}
                         />
                         {errors.name && (
-                          <div className="mt-1 text-sm text-red-400 invalid-feedback">
+                          <div className="invalid-feedback mt-1 text-sm text-red-400">
                             {errors.name.message}
                           </div>
                         )}
@@ -175,7 +189,7 @@ export default function PopupWidget() {
                       <div className="mb-4">
                         <label
                           htmlFor="email"
-                          className="block mb-2 text-sm text-trueZinc-600 dark:text-trueZinc-400"
+                          className="mb-2 block text-sm text-trueZinc-700"
                         >
                           Email Address
                         </label>
@@ -190,15 +204,15 @@ export default function PopupWidget() {
                             },
                           })}
                           placeholder="you@company.com"
-                          className={`w-full px-3 py-2 placeholder-trueZinc-300 bg-trueZinc-100 border border-trueZinc-300 rounded-md focus:outline-none focus:ring   ${
+                          className={`w-full rounded-md border border-trueZinc-300 bg-trueZinc-100 px-3 py-2 placeholder-trueZinc-300 focus:outline-none focus:ring   ${
                             errors.email
-                              ? "border-red-600 focus:border-red-600 ring-red-100"
-                              : "border-trueZinc-300 focus:border-truePurple-600 ring-truePurple-100"
+                              ? "border-red-600 ring-red-100 focus:border-red-600"
+                              : "border-trueZinc-300 ring-trueFushia-100 focus:border-trueFushia-600"
                           }`}
                         />
 
                         {errors.email && (
-                          <div className="mt-1 text-sm text-red-400 invalid-feedback">
+                          <div className="invalid-feedback mt-1 text-sm text-red-400">
                             {errors.email.message}
                           </div>
                         )}
@@ -207,7 +221,7 @@ export default function PopupWidget() {
                       <div className="mb-4">
                         <label
                           htmlFor="message"
-                          className="block mb-2 text-sm text-trueZinc-600 dark:text-trueZinc-400"
+                          className="mb-2 block text-sm text-trueZinc-700"
                         >
                           Your Message
                         </label>
@@ -219,15 +233,15 @@ export default function PopupWidget() {
                             required: "Enter your Message",
                           })}
                           placeholder="Your Message"
-                          className={`w-full px-3 py-2 placeholder-trueZinc-300 bg-trueZinc-100 border border-trueZinc-300 rounded-md h-28 focus:outline-none focus:ring   ${
+                          className={`h-28 w-full rounded-md border border-trueZinc-300 bg-trueZinc-100 px-3 py-2 placeholder-trueZinc-300 focus:outline-none focus:ring   ${
                             errors.message
-                              ? "border-red-600 focus:border-red-600 ring-red-100"
-                              : "border-trueZinc-300 focus:border-truePurple-600 ring-truePurple-100"
+                              ? "border-red-600 ring-red-100 focus:border-red-600"
+                              : "border-trueZinc-300 ring-trueFushia-100 focus:border-trueFushia-600"
                           }`}
                           required
                         ></textarea>
                         {errors.message && (
-                          <div className="mt-1 text-sm text-red-400 invalid-feedback">
+                          <div className="invalid-feedback mt-1 text-sm text-red-400">
                             {errors.message.message}
                           </div>
                         )}
@@ -235,11 +249,11 @@ export default function PopupWidget() {
                       <div className="mb-3">
                         <button
                           type="submit"
-                          className="w-full px-3 py-4 text-trueZinc-100 bg-truePurple-700 rounded-md focus:bg-trueSky-600 focus:outline-none"
+                          className="w-full rounded-md bg-trueFushia-700 px-3 py-4 text-trueZinc-100 focus:bg-truePink-600 focus:outline-none"
                         >
                           {isSubmitting ? (
                             <svg
-                              className="w-5 h-5 mx-auto text-trueZinc-100 animate-spin"
+                              className="mx-auto h-5 w-5 animate-spin text-trueZinc-100"
                               xmlns="http://www.w3.org/2000/svg"
                               fill="none"
                               viewBox="0 0 24 24"
@@ -268,21 +282,13 @@ export default function PopupWidget() {
 
                   {isSubmitSuccessful && isSuccess && (
                     <>
-                      <div className="flex flex-col items-center justify-center h-full text-center text-trueZinc-100 rounded-md">
-                        <svg
-                          width="60"
-                          height="60"
-                          className="text-green-300"
-                          viewBox="0 0 100 100"
-                          fill="none"
-                          xmlns="http://www.w3.org/2000/svg"
-                        >
-                          <path
-                            d="M26.6666 50L46.6666 66.6667L73.3333 33.3333M50 96.6667C43.8716 96.6667 37.8033 95.4596 32.1414 93.1144C26.4796 90.7692 21.3351 87.3317 17.0017 82.9983C12.6683 78.6649 9.23082 73.5204 6.8856 67.8586C4.54038 62.1967 3.33331 56.1283 3.33331 50C3.33331 43.8716 4.54038 37.8033 6.8856 32.1414C9.23082 26.4796 12.6683 21.3351 17.0017 17.0017C21.3351 12.6683 26.4796 9.23084 32.1414 6.88562C37.8033 4.5404 43.8716 3.33333 50 3.33333C62.3767 3.33333 74.2466 8.24998 82.9983 17.0017C91.75 25.7534 96.6666 37.6232 96.6666 50C96.6666 62.3768 91.75 74.2466 82.9983 82.9983C74.2466 91.75 62.3767 96.6667 50 96.6667Z"
-                            stroke="currentColor"
-                            strokeWidth="3"
-                          />
-                        </svg>
+                      <div className="flex h-full flex-col items-center justify-center rounded-md text-center text-trueZinc-100">
+                        <Player
+                          autoplay
+                          loop
+                          src="https://storage.fleek.zone/ead18a13-670a-4701-91a7-06a767c3dcd5-bucket/assets/lottie/drunkbeerbottlegreen.json"
+                          style={{ height: "100px", width: "100px" }}
+                        ></Player>
                         <h3 className="py-5 text-xl text-green-500">
                           Message sent successfully
                         </h3>
@@ -308,28 +314,20 @@ export default function PopupWidget() {
                   )}
 
                   {isSubmitSuccessful && !isSuccess && (
-                    <div className="flex flex-col items-center justify-center h-full text-center text-trueZinc-100 rounded-md">
-                      <svg
-                        width="60"
-                        height="60"
-                        viewBox="0 0 97 97"
-                        className="text-red-400"
-                        fill="none"
-                        xmlns="http://www.w3.org/2000/svg"
-                      >
-                        <path
-                          d="M27.9995 69C43.6205 53.379 52.3786 44.621 67.9995 29M26.8077 29L67.9995 69M48.2189 95C42.0906 95 36.0222 93.7929 30.3604 91.4477C24.6985 89.1025 19.554 85.6651 15.2206 81.3316C10.8872 76.9982 7.44975 71.8538 5.10454 66.1919C2.75932 60.53 1.55225 54.4617 1.55225 48.3333C1.55225 42.205 2.75932 36.1366 5.10454 30.4748C7.44975 24.8129 10.8872 19.6684 15.2206 15.335C19.554 11.0016 24.6985 7.56418 30.3604 5.21896C36.0222 2.87374 42.0906 1.66667 48.2189 1.66667C60.5957 1.66667 72.4655 6.58333 81.2172 15.335C89.9689 24.0867 94.8856 35.9566 94.8856 48.3333C94.8856 60.7101 89.9689 72.58 81.2172 81.3316C72.4655 90.0833 60.5957 95 48.2189 95Z"
-                          stroke="CurrentColor"
-                          strokeWidth="3"
-                        />
-                      </svg>
+                    <div className="flex h-full flex-col items-center justify-center rounded-md text-center text-trueZinc-100">
+                      <Player
+                        autoplay
+                        loop
+                        src="https://storage.fleek.zone/ead18a13-670a-4701-91a7-06a767c3dcd5-bucket/assets/lottie/drunkbeerbottlered.json"
+                        style={{ height: "100px", width: "100px" }}
+                      ></Player>
 
-                      <h3 className="text-xl text-red-400 py-7">
+                      <h3 className="py-7 text-xl text-red-400">
                         Oops, Something went wrong!
                       </h3>
                       <p className="text-trueZinc-700 md:px-3">{Message}</p>
                       <button
-                        className="mt-6 text-truePurple-600 focus:outline-none"
+                        className="mt-6 text-trueFushia-600 focus:outline-none"
                         onClick={() => reset()}
                       >
                         Go back
