@@ -1,8 +1,10 @@
+'use client';
 import Head from "next/head";
+import Script from "next/script";
 
 export default function SEO({
-  title = "0xaloysius",
-  description = "Entrepreneur in technology, biotech, cleantech, foodtech & mobility.",
+  title = "4mobeers",
+  description = "Home of the Road Kingz",
 }) {
   return (
     <>
@@ -11,27 +13,21 @@ export default function SEO({
         <meta key="description" name="description" content={description} />
 
         <meta property="og:type" content="website" />
-        <meta property="og:url" content="https://0xaloysius.x" />
-        <meta property="og:title" content="0xaloysius" />
-        <meta
-          property="og:description"
-          content="Entrepreneur in technology, biotech, cleantech, foodtech & mobility."
-        />
+        <meta property="og:url" content="https://4mobeers.nft" />
+        <meta property="og:title" content="4mobeers" />
+        <meta property="og:description" content="4MoBeers" />
         <meta
           property="og:image"
-          content="https://storageapi.fleek.co/47c43c10-6c0c-4b3b-a520-683839c214cf-bucket/assets/img/0xaloysius-purple.svg"
+          content="https://storageapi.fleek.co/47c43c10-6c0c-4b3b-a520-683839c214cf-bucket/assets/img/4mobeers-purple.svg"
         />
 
         <meta property="twitter:card" content="summary_large_image" />
-        <meta property="twitter:url" content="https://0xaloysius.x" />
-        <meta property="twitter:title" content="0xaloysius" />
-        <meta
-          property="twitter:description"
-          content="Entrepreneur in technology, biotech, cleantech, foodtech & mobility."
-        />
+        <meta property="twitter:url" content="https://4mobeers.nft" />
+        <meta property="twitter:title" content="4mobeers" />
+        <meta property="twitter:description" content="4MoBeers" />
         <meta
           property="twitter:image"
-          content="https://storageapi.fleek.co/47c43c10-6c0c-4b3b-a520-683839c214cf-bucket/assets/img/0xaloysius-purple.svg"
+          content="https://storageapi.fleek.co/47c43c10-6c0c-4b3b-a520-683839c214cf-bucket/assets/img/4mobeers-purple.svg"
         />
       </Head>
     </>
