@@ -1,3 +1,4 @@
+import prettier from "prettier";
 import Image from "next/image";
 import React from "react";
 import Container from "./container";
@@ -11,7 +12,7 @@ export default function Testimonials() {
     <Container>
       <div className="grid gap-10 lg:grid-cols-2 xl:grid-cols-3">
         <div className="lg:col-span-2 xl:col-auto">
-          <div className="flex flex-col justify-between w-full h-full bg-trueZinc-100 px-14 rounded-2xl py-14 dark:bg-trueZinc-800">
+          <div className="flex h-full w-full flex-col justify-between rounded-2xl bg-trueZinc-100 px-14 py-14 dark:bg-trueZinc-800">
             <p className="text-2xl leading-normal ">
               Share a real <Mark>testimonial</Mark>
               that hits some of your benefits from one of your popular customer.
@@ -25,7 +26,7 @@ export default function Testimonials() {
           </div>
         </div>
         <div className="">
-          <div className="flex flex-col justify-between w-full h-full bg-trueZinc-100 px-14 rounded-2xl py-14 dark:bg-trueZinc-800">
+          <div className="flex h-full w-full flex-col justify-between rounded-2xl bg-trueZinc-100 px-14 py-14 dark:bg-trueZinc-800">
             <p className="text-2xl leading-normal ">
               Make sure you only pick the <Mark>right sentence</Mark>
               to keep it short and simple.
@@ -39,7 +40,7 @@ export default function Testimonials() {
           </div>
         </div>
         <div className="">
-          <div className="flex flex-col justify-between w-full h-full bg-trueZinc-100 px-14 rounded-2xl py-14 dark:bg-trueZinc-800">
+          <div className="flex h-full w-full flex-col justify-between rounded-2xl bg-trueZinc-100 px-14 py-14 dark:bg-trueZinc-800">
             <p className="text-2xl leading-normal ">
               This is an <Mark>awesome</Mark> landing page template I&apos;ve
               seen. I would use this for anything.
@@ -59,8 +60,8 @@ export default function Testimonials() {
 
 function Avatar(props) {
   return (
-    <div className="flex items-center mt-8 space-x-3">
-      <div className="flex-shrink-0 overflow-hidden rounded-full w-14 h-14">
+    <div className="mt-8 flex items-center space-x-3">
+      <div className="h-14 w-14 flex-shrink-0 overflow-hidden rounded-full">
         <Image
           src={props.image}
           width="40"
@@ -84,7 +85,7 @@ function Mark(props) {
   return (
     <>
       {" "}
-      <mark className="text-trueZinc-800 bg-trueSky-100 rounded-md ring-trueSky-100 ring-4 dark:ring-trueSky-900 dark:bg-trueSky-900 dark:text-trueSky-200">
+      <mark className="rounded-md bg-trueSky-100 text-trueZinc-800 ring-4 ring-trueSky-100 dark:bg-trueSky-900 dark:text-trueSky-200 dark:ring-trueSky-900">
         {props.children}
       </mark>{" "}
     </>
