@@ -1,6 +1,8 @@
-/** @type {import('next').NextConfig} */
+/* @type {import('next').NextConfig} */
 const nextConfig = {
   reactStrictMode: true,
+  basePath: "",
+  pageExtensions: ["js", "jsx", "ts", "tsx", "mdx"],
   images: {
     //domains: ["storageapi.fleek.co", "savivets.wpengine.com"],
     remotePatterns: [
