@@ -1,21 +1,9 @@
 import React from "react";
 import Document, { Html, Head, Main, NextScript } from "next/document";
 
-import Script from "next/script";
-import Navbar from "../components/navbar";
-import Footer from "../components/footermini";
-
 //import Head from "next/head";
 
 class MyDocument extends Document {
-  static async getInitialProps(ctx) {
-    const initialProps = await Document.getInitialProps(ctx);
-    return {
-      ...initialProps,
-      styles: React.Children.toArray([initialProps.styles]),
-    };
-  }
-
   render() {
     return (
       <Html>
@@ -30,10 +18,7 @@ class MyDocument extends Document {
           />
 
           <meta property="twitter:card" content="summary_large_image" />
-          <meta
-            property="twitter:url"
-            content="https://twitter.com/4mobeers"
-          />
+          <meta property="twitter:url" content="https://twitter.com/4mobeers" />
           <meta property="twitter:title" content="4mobeers" />
           <meta
             property="twitter:description"
@@ -45,7 +30,7 @@ class MyDocument extends Document {
           />
           <link rel="me" href="https://noc.social/@4mobeers" />
           <link rel="preconnect" href="https://fonts.googleapis.com" />
-          <link rel="preconnect" href="https://fonts.gstatic.com" crossOrigin />
+          <link rel="preconnect" href="https://fonts.gstatic.com" />
           <link
             href="https://fonts.googleapis.com/css2?family=Sora:wght@700&display=swap"
             rel="stylesheet"
@@ -54,7 +39,6 @@ class MyDocument extends Document {
             rel="icon"
             href="https://storageapi.fleek.one/ead18a13-670a-4701-91a7-06a767c3dcd5-bucket/assets/img/logo-4mobeers.svg"
           />
-          
         </Head>
         <body className="lightmode dark:darkmode">
           <Main />
