@@ -3,7 +3,7 @@ const colors = require("tailwindcss/colors");
 
 module.exports = {
   mode: "jit",
-  content: ["./pages/**/*.{html,js}", "./components/**/*.{html,js}"],
+  content: ["./pages/**/*.{html,js}", "./components/**/*.{html,js,jsx}"],
   darkMode: "class", // or 'media' or 'class'
 
   backgroundImage: {
@@ -17,8 +17,18 @@ module.exports = {
         "gradient-1": "animate-gradient-1 5s infinite",
         "gradient-2": "animate-gradient-2 5s infinite",
         "gradient-3": "animate-gradient-3 5s infinite",
+        linear: "backgroundLinear 3s linear infinite",
+        wiggle: "4s linear infinite wiggle",
       },
       keyframes: {
+        wiggle: {
+          "0%, 100%": { transform: "rotate(-5deg)" },
+          "50%": { transform: "rotate(20deg)" },
+        },
+        backgroundLinear: {
+          "0%": { backgroundPosition: "0% 50%" },
+          "100%": { backgroundPosition: "200% 50%" },
+        },
         "animate-gradient-1": {
           "0%, 16.667%, 100%": { opacity: "1" },
           "33.333%, 83.333%": { opacity: "0" },
@@ -43,6 +53,7 @@ module.exports = {
         trueEmerald: colors.emerald,
         trueBlue: colors.blue,
         truePink: colors.pink,
+        trueFushia: colors.fuchsia,
       },
     },
     fontFamily: {
@@ -53,5 +64,7 @@ module.exports = {
   variants: {
     extend: {},
   },
-  plugins: [require("@tailwindcss/aspect-ratio")],
+  plugins: [
+    require("@tailwindcss/aspect-ratio", "prettier-plugin-tailwindcss"),
+  ],
 };
